refactor: BaseEntity 패키지 이동
