+++ src/main/kotlin/com/example/mykku/dailymessage/domain/DailyMessage.kt
@@ -1,7 +1,8 @@
 package com.example.mykku.dailymessage.domain
 
-import com.example.mykku.common.BaseEntity
+import com.example.mykku.common.domain.BaseEntity
 import jakarta.persistence.*
+import java.time.LocalDate
 
 @Entity
 class DailyMessage(
@@ -11,6 +12,9 @@ class DailyMessage(
 
     @Column(name = "content", nullable = false)
     var content: String,
+
+    @Column(name = "date", nullable = false)
+    var date: LocalDate,
 
     @OneToMany(mappedBy = "dailyMessage", cascade = [CascadeType.ALL], orphanRemoval = true, fetch = FetchType.LAZY)
     var comments: MutableList<DailyMessageComment> = mutableListOf(),
