+++ src/main/kotlin/com/example/mykku/feed/domain/FeedImage.kt
@@ -1,6 +1,6 @@
 package com.example.mykku.feed.domain
 
-import com.example.mykku.common.BaseEntity
+import com.example.mykku.common.domain.BaseEntity
 import jakarta.persistence.*
 
 @Entity
