+++ src/main/kotlin/com/example/mykku/member/domain/LikeFeedComment.kt
@@ -1,6 +1,6 @@
 package com.example.mykku.member.domain
 
-import com.example.mykku.common.BaseEntity
+import com.example.mykku.common.domain.BaseEntity
 import com.example.mykku.feed.domain.FeedComment
 import jakarta.persistence.*
