+++ src/main/kotlin/com/example/mykku/member/domain/SaveDailyMessage.kt
@@ -1,6 +1,6 @@
 package com.example.mykku.member.domain
 
-import com.example.mykku.common.BaseEntity
+import com.example.mykku.common.domain.BaseEntity
 import com.example.mykku.dailymessage.domain.DailyMessage
 import jakarta.persistence.*
